+++ pkg/replication/metrics.go
@@ -1,0 +1,50 @@
+//
+// DISCLAIMER
+//
+// Copyright 2023 ArangoDB GmbH, Cologne, Germany
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// Copyright holder is ArangoDB GmbH, Cologne, Germany
+//
+
+package replication
+
+import (
+	"github.com/arangodb/kube-arangodb/pkg/generated/metric_descriptions"
+	"github.com/arangodb/kube-arangodb/pkg/util/metrics"
+)
+
+type Metrics struct {
+	DeploymentReplication struct {
+		Active, Failed bool
+	}
+}
+
+func (dr *DeploymentReplication) CollectMetrics(m metrics.PushMetric) {
+	name, namespace := dr.apiObject.GetName(), dr.apiObject.GetNamespace()
+
+	toGauge := func(b bool) float64 {
+		if b {
+			return 1
+		}
+		return 0
+	}
+
+	m.Push(metric_descriptions.ArangodbOperatorResourcesArangodeploymentreplicationActiveGauge(
+		toGauge(dr.metrics.DeploymentReplication.Active), namespace, name),
+	)
+	m.Push(metric_descriptions.ArangodbOperatorResourcesArangodeploymentreplicationFailedGauge(
+		toGauge(dr.metrics.DeploymentReplication.Failed), namespace, name),
+	)
+}