+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 - (Maintenance) Bump K8S Version to 1.24.16
 - (Maintenance) Bump K8S Version to 1.25.12
 - (Maintenance) Bump Go to 1.20.7
+- (Feature) Add basic metrics for ArangoDeploymentReplication CR 
 
 ## [1.2.31](https://github.com/arangodb/kube-arangodb/tree/1.2.31) (2023-07-14)
 - (Improvement) Block traffic on the services if there is more than 1 active leader in ActiveFailover mode